Commented TSBSGeant4File class. I take advantage of that to explain a few of the modifications this class has been subject to.
A function to calculate the range in the GEM gas for very low energy electron has been added.

This is necessary as we do not have, in g4sbs, the information of the track position leaving
the sensitive gas (although on second thought it might not be too complicated to add it.),
making it necessary to calculate it by hand.

If the energy is very low, the momentum slopes (dpx/dpz, dpy/dpz) are often very high,
which implies a fantaisist value for the position of the track leaving the sensitive gas.
The range calculation prevents that, as it will set a sensical limit to the path of very low electrons.