Added the example macro DigDemo3 in the example section. With this macro one obtained a file with digitized GEM hits. The digitized data stored in the output file is compatible with the data stored in the input file (at least the MC truth data are the same; this has been tested on a one event).
Along the way, a few fixes have been made.
For instance, the confusion between Particle Type (actually the PDG PID)
and the G4 track ID has been disentangled in classes TSBSGeant4File, TSolEVIOFile, and TSolGEMData.
(for instance, this data was sometimes declared as an unsigned int,
which is problematic when the PID is negative).

Also fixed the few lines of codes in TSBSGeant4File which determined
if a MC track was stored already or not (this used to stroe duplicate tracks).

The modifications in TSBSSimGEMDigitization.cxx only concern printouts which have been commented out.