Modified class TSBSSimGEMDigitization (and TSolSimGEMDig...) to update the digitization. Digitization refinement has been performed by Weizhi, who kindly shared his changes at: https://github.com/xweizhi/libsolgem.
It includes, among others:
 - refinment of the avalnche model;
 - inclusion of the cross-talk;
 - inclusion of the pedestal.
More details can be found at:
http://hallaweb.jlab.org/12GeV/SoLID/meeting_coll/2016_12/SoLIDTracking_Weizhi_12022016.pdf

I also incorporated the convenient class TSolDBManager, which is again credited to Weizhi
(although I still need to adapt it to SBS).

Finally, I have added in the repository a few new data files,
and deleted a bunch of files and symlinks that have become useless.