Don't return const& from simple getters. Instead, make the getters const.
Some of the Get functions, e.g. GetDMag, cannot easily be made const
because they use map lookups via operator[] internally, which is not
a const operation. operator[] creates default-constructed map entries
if the requested key does not yet exist in the map. I wonder if the
code relies on this auto-create behavior or if there is a
misunderstanding of how std::maps work. If the latter, the operator[]
lookups are actually latent bugs and should be replaced with find().