At present, libsbsgem does not actually depend on EVIO at all
The C++ API is not used in TSBSGeant4File anymore. Instead, this
class reads a ROOT file.

TSBSSimFile includes evio.h, but only for the definition of the
return code S_SUCCESS, which an implementation of THaRunBase shouldn't
use anyway. Changed TSBSSimFile to use the correct return codes
READ_* (see THaRunBase.h)